+++ joss/paper.md
@@ -15,7 +15,7 @@ authors:
     orcid: 0009-0005-0148-2499
     affiliation: 1
 affiliations:
-  - name: University of California, Berkeley, Berkeley, CA 94720, USA
+  - name: University of California, Berkeley, Berkeley, CA 94720, United States of America
     index: 1
 date: 21 April 2024
 bibliography: paper.bib
@@ -45,7 +45,7 @@ APackOfTheClones also offers a suite of methods for visualizing and analyzing si
 
 The main clonal expansion visualization that APackOfTheClones implements is shown in the following figure \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
 
-![A single-cell experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.svg)
+![A single-cell experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
 
 The visualization gives the immediate insight that certain cell subsets such as those in cluster four contains many more expanded clones both by quantity and frequency.
 
