+++ joss/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={103--112},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.imlet.2014.10.011}
 }
 
 @article{huang2022role,
@@ -17,7 +18,8 @@
   number={5},
   pages={629--634},
   year={2022},
-  publisher={American College of Cardiology Foundation Washington DC}
+  publisher={American College of Cardiology Foundation Washington DC},
+  doi={10.1016/j.jaccao.2022.08.012}
 }
 
 @article{hao2023dictionary,
@@ -26,7 +28,8 @@
   journal={Nature biotechnology},
   pages={1--12},
   year={2023},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41587-023-01767-y}
 }
 
 @article{ma2021single,
@@ -37,7 +40,8 @@
   number={62},
   pages={eabg6356},
   year={2021},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/sciimmunol.abg6356}
 }
 
 @article{stark2022recombinant,
@@ -48,7 +52,8 @@
   number={10},
   pages={2987--3001},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/all.15399}
 }
 
 @article{den2014activation,
@@ -59,7 +64,8 @@
   number={2},
   pages={103--112},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.3389/fimmu.2019.00360}
 }
 
 @article{sturm2020scirpy,
@@ -70,7 +76,8 @@
   number={18},
   pages={4817--4818},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.37473/dac/10.1101/2020.04.10.035865}
 }
 
 @article{huang2022role,
@@ -81,16 +88,8 @@
   number={5},
   pages={629--634},
   year={2022},
-  publisher={American College of Cardiology Foundation Washington DC}
-}
-
-@article{hao2023dictionary,
-  title={Dictionary learning for integrative, multimodal and scalable single-cell analysis},
-  author={Hao, Yuhan and Stuart, Tim and Kowalski, Madeline H and Choudhary, Saket and Hoffman, Paul and Hartman, Austin and Srivastava, Avi and Molla, Gesmira and Madad, Shaista and Fernandez-Granda, Carlos and others},
-  journal={Nature biotechnology},
-  pages={1--12},
-  year={2023},
-  publisher={Nature Publishing Group US New York}
+  publisher={American College of Cardiology Foundation Washington DC},
+  doi={10.1016/j.jaccao.2022.08.012}
 }
 
 @article{andrews2021tutorial,
@@ -101,49 +100,15 @@
   number={1},
   pages={1--9},
   year={2021},
-  publisher={Nature Publishing Group UK London}
-}
-
-@article{ma2021single,
-  title={Single-cell analysis pinpoints distinct populations of cytotoxic CD4+ T cells and an IL-10+ CD109+ TH2 cell population in nasal polyps},
-  author={Ma, Junjie and Tibbitt, Christopher A and Geor{\'e}n, Susanna Kumlien and Christian, Murray and Murrell, Ben and Cardell, Lars-Olaf and Bachert, Claus and Coquet, Jonathan M},
-  journal={Science Immunology},
-  volume={6},
-  number={62},
-  pages={eabg6356},
-  year={2021},
[…]
