### 2024-06-23 #6868 @Qile0317
@jspaezp DOIs are now all fixed where availiable - lemme know if there's anything else I could help with to speed up the process
