### 2024-06-26 Qile0317/scballpack#10 @ncborcherding
JOSS Review Manuscript Editing
Hey @Qile0317

Can you go through the manuscript and just grammar/spell check things? I am finding just small issues, but that is one of the reviewer criterias

Example:

Nick
