### 2024-06-26 #6868 @ncborcherding
@Qile0317 Grammar/Quality of Writing issue per below issue.

https://github.com/Qile0317/APackOfTheClones/issues/10#issue-2375632966
