### 2024-06-28 0c91b91
spelling fixes and added ORCID into DESCRIPTION
