### 2024-06-28 13f06c7
spelling fixes and added ORCID into DESCRIPTION
