+++ joss/paper.md
@@ -31,11 +31,11 @@ The R [@r2023r] package APackOfTheClones implements and *extends* a novel method
 
 # Statement of need
 
-Cells in single-cell experiments are conventionally visualized on a reduction scatterplot where each point represents a cell after all its features have been projected into two dimensions. The points can be colored by different factors to display useful information, including its asssigned identity. [@andrews2021tutorial]
+Cells in single-cell experiments are conventionally visualized on a reduction scatterplot where each point represents a cell after all its features have been projected into two dimensions. The points can be colored by different factors to display useful information, including its assigned identity. [@andrews2021tutorial]
 
-Understanding the role of specific cells in various contexts requires understanding of relationships between cell populations and their behaviors. One attribute is the clonal expansion dynamics, which are inferred from downstream analyses. [@adams2020clonal] Overlaying clonal expansion information on a by-clonotype basis onto the reduction scatterplot of the cells of an experiment aids in adding an additional layer of insight, allowing for a swift, intuitive understanding of how clonal dynamics relate to the identified cellular subsets. For example, it can help gauge the presence of hyperexpanded clones for each cell type; compare potential changes in frequencies after certain theraputic treatments, etc.
+Understanding the role of specific cells in various contexts requires understanding of relationships between cell populations and their behaviors. One attribute is the clonal expansion dynamics, which are inferred from downstream analyses. [@adams2020clonal] Overlaying clonal expansion information on a by-clonotype basis onto the reduction scatterplot of the cells of an experiment aids in adding an additional layer of insight, allowing for a swift, intuitive understanding of how clonal dynamics relate to the identified cellular subsets. For example, it can help gauge the presence of hyper expanded clones for each cell type; compare potential changes in frequencies after certain therapeutic treatments, etc.
 
-There is no standardized convention to visualize this single cell level clonotype data on the dimensional reduction. Some of the current approaches include 1. Using a color gradient corresponding to each frequency to highlight each individual point by the clonal expansion, implpemented in *scRepertoire* and *scirpy* [sturm2020scirpy] 2. Overlaying a 2D contour where points representing clones with higher frequencies have elevated levels [@andreatta2023tcell] 3. Increasing sizes of points based on clonal frequency, used in figure 2c of @wang2021single.
+There is no standardized convention to visualize this single cell level clonotype data on the dimensional reduction. Some of the current approaches include 1. Using a color gradient corresponding to each frequency to highlight each individual point by the clonal expansion, implemented in *scRepertoire* and *scirpy* [sturm2020scirpy] 2. Overlaying a 2D contour where points representing clones with higher frequencies have elevated levels [@andreatta2023tcell] 3. Increasing sizes of points based on clonal frequency, used in figure 2c of @wang2021single.
 
 These approaches capture approximate, global trends but do not provide an exact representation. They lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. From a visual standpoint, This limitation hinders subtle variations in clonal expansion patterns. APackOfTheClones solves this issue by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses.
 
@@ -49,16 +49,16 @@ The main clonal expansion visualization that APackOfTheClones implements is show
 
 The visualization gives the immediate insight that certain cell subsets such as those in cluster four contains many more expanded clones both by quantity and frequency.
 
[…]
