### 2024-06-28 Qile0317/scballpack#10 @Qile0317
Just used cspell to fix all the spelling mistakes it could spot on the joss branch in `joss/paper.md`, and after looking through most of the writing I believe the major issues are all fixed. Closing and continuing the discussion on the JOSS review.
