### 2024-06-28 #6868 @Qile0317
Grammar issues have been fixed to the best of my efforts.

A question to both @ncborcherding and @jspaezp - if you check the [comments on the sourcecode of my paper](https://github.com/Qile0317/APackOfTheClones/blob/13f06c7d29f0940d3b3f8864bae81ec2bf73f03f/joss/paper.md?plain=1#L54C1-L58C1) - would you think it is nessecary to have further discussions on the other features of the package? Though it may not be the most useful, there are functions for getting and filtering clone sizes, highlighting clonotypes, and adding links between shared clones.
