### 2024-06-28 8094dab
spelling fixes, rm deprecation startup msg, fixed description, rm magrittr dep
