+++ R/plot_API.R
@@ -153,6 +153,6 @@ get_retain_scale_dims <- function(
 
 check_is_apotc_ggplot <- function(x) {
   if (!isApotcGGPlot(x)) {
-    stop(call. = FALSE, "not an output of `APOTCPlot` or `vizAPOTC`")
+    stop(call. = FALSE, "not an output of `APOTCPlot()` or `vizAPOTC()`")
   }
 }
