+++ vignettes/APackOfTheClones-shared.Rmd
@@ -30,7 +30,7 @@ quiet_load_all_CRAN <- function(...) {
 }
 
 # load packages
-quiet_load_all_CRAN("ggplot2", "Seurat", "magrittr", "APackOfTheClones")
+quiet_load_all_CRAN("ggplot2", "Seurat", "dplyr", "APackOfTheClones")
 
 # load data
 pbmc <- get(data("combined_pbmc"))
@@ -40,7 +40,7 @@ pbmc <- get(data("combined_pbmc"))
 
 A novel feature in this package is the ability to get, filter, and visualize clonotypes that are common across seurat clusters. Potential applications include the ability to quickly and intuitively gauge the rough clonal heterogeneity of certain clusters. Note that in scRepertoire, the `scRepertoire::clonalNetwork` function does something similar for the original dimensional reductions using arrows, though those display the aggregated degree of commonality between each cluster.
 
-Please read `vignette("APackOfTheClones")` and ideally also `vignette("APackOfTheClones-runs")` before reading this vignette. One should then also see that although some of the visualizations in this vignette were produced with the `vizAPOTC` function, all visualizations can be replicated with `RunAPOTC` and `APOTCPlot` that both share parts of the arguments of `vizAPOTC`.
+Please read `vignette("APackOfTheClones")` and ideally also `vignette("APackOfTheClones-runs")` before reading this vignette. One should then also see that although some of the visualizations in this vignette were produced with the `vizAPOTC()` function, all visualizations can be replicated with `RunAPOTC()` and `APOTCPlot()` that both share parts of the arguments of `vizAPOTC()`.
 
 ## Getting Clonotypes Common Across Seurat Clusters
 
