+++ man/AdjustAPOTC.Rd
@@ -69,7 +69,7 @@ could pass in something like \code{extra_filter = "nchar(CTaa) - 1 > 10"}. When
 involving characters, ensure to enclose with single quotes.}
 
 \item{run_id}{character. This will be the ID associated with the data of a
-run, and will be used by other important functions like \link{APOTCPlot} and
+run, and will be used by other important functions like \code{\link[=APOTCPlot]{APOTCPlot()}} and
 \link{AdjustAPOTC}. Defaults to \code{NULL}, in which case the ID will be generated
 in the following format:
 
@@ -160,9 +160,9 @@ The adjusted \code{seurat_obj}
 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
 
 If the user is unsatisfied with the clonal expansion plot that
-was generated from \code{RunAPOTC} and \code{APOTCPlot}, this function has a range of
+was generated from \code{\link[=RunAPOTC]{RunAPOTC()}} and \code{\link[=APOTCPlot]{APOTCPlot()}}, this function has a range of
 arguments to modify the data and/or parameters of the visualization. Note
-that some of the arguments may conflict with eachother.
+that some of the arguments may conflict with each other.
 }
 \examples{
 # do an APackOfTheClones run
