### 2024-06-28 732950b
spelling fixes, rm deprecation startup msg, fixed description, rm magrittr dep
