### 2024-06-29 12d729b
modified README and NEWS
