### 2024-06-29 f348f15
modified README and NEWS
