### 2024-06-29 681db83
delete joss and cran comments for v1.2.1
