+++ joss/paper.bib
@@ -1,206 +1,0 @@
-@article{den2014activation,
-  title={The activation of the adaptive immune system: cross-talk between antigen-presenting cells, T cells and B cells},
-  author={den Haan, Joke MM and Arens, Ramon and van Zelm, Menno C},
-  journal={Immunology letters},
-  volume={162},
-  number={2},
-  pages={103--112},
-  year={2014},
-  publisher={Elsevier},
-  doi={10.1016/j.imlet.2014.10.011}
-}
-
-@article{huang2022role,
-  title={The role of single-cell profiling and deep immunophenotyping in understanding immune therapy cardiotoxicity},
-  author={Huang, Yuhsin Vivian and Waliany, Sarah and Lee, Daniel and Galdos, Francisco X and Witteles, Ronald M and Neal, Joel W and Fan, Alice C and Maecker, Holden T and Nguyen, Patricia K and Wu, Sean M and others},
-  journal={Cardio Oncology},
-  volume={4},
-  number={5},
-  pages={629--634},
-  year={2022},
-  publisher={American College of Cardiology Foundation Washington DC},
-  doi={10.1016/j.jaccao.2022.08.012}
-}
-
-@article{hao2023dictionary,
-  title={Dictionary learning for integrative, multimodal and scalable single-cell analysis},
-  author={Hao, Yuhan and Stuart, Tim and Kowalski, Madeline H and Choudhary, Saket and Hoffman, Paul and Hartman, Austin and Srivastava, Avi and Molla, Gesmira and Madad, Shaista and Fernandez-Granda, Carlos and others},
-  journal={Nature biotechnology},
-  pages={1--12},
-  year={2023},
-  publisher={Nature Publishing Group US New York},
-  doi={10.1038/s41587-023-01767-y}
-}
-
-@article{ma2021single,
-  title={Single-cell analysis pinpoints distinct populations of cytotoxic CD4+ T cells and an IL-10+ CD109+ TH2 cell population in nasal polyps},
-  author={Ma, Junjie and Tibbitt, Christopher A and Geor{\'e}n, Susanna Kumlien and Christian, Murray and Murrell, Ben and Cardell, Lars-Olaf and Bachert, Claus and Coquet, Jonathan M},
-  journal={Science Immunology},
-  volume={6},
-  number={62},
-  pages={eabg6356},
-  year={2021},
-  publisher={American Association for the Advancement of Science},
-  doi={10.1126/sciimmunol.abg6356}
-}
-
-@article{stark2022recombinant,
-  title={Recombinant multimeric dog allergen prevents airway hyperresponsiveness in a model of asthma marked by vigorous TH2 and TH17 cell responses},
-  author={Stark, Julian M and Liu, Jielu and Tibbitt, Christopher A and Christian, Murray and Ma, Junjie and Wintersand, Anna and Dunst, Josefine and Kreslavsky, Taras and Murrell, Ben and Adner, Mikael and others},
-  journal={Allergy},
-  volume={77},
-  number={10},
-  pages={2987--3001},
-  year={2022},
-  publisher={Wiley Online Library},
-  doi={10.1111/all.15399}
-}
-
-@article{den2014activation,
-  title={The activation of the adaptive immune system: cross-talk between antigen-presenting cells, T cells and B cells},
-  author={den Haan, Joke MM and Arens, Ramon and van Zelm, Menno C},
-  journal={Immunology letters},
-  volume={162},
-  number={2},
-  pages={103--112},
-  year={2014},
-  publisher={Elsevier},
-  doi={10.3389/fimmu.2019.00360}
-}
-
[…]
