### 2024-06-29 bf80477
delete joss and cran comments for v1.2.1
