+++ joss/paper.md
@@ -1,67 +1,0 @@
----
-title: 'APackOfTheClones: Visualization of clonal expansion with circle packing'
-tags:
-  - R
-  - dimensional reduction
-  - clonal expansion
-  - immunology
-  - immunoinformatics
-  - bioinformatics
-  - computational biology
-  - seurat
-  - single cell
-authors:
-  - name: Qile Yang
-    orcid: 0009-0005-0148-2499
-    affiliation: 1
-affiliations:
-  - name: University of California, Berkeley, Berkeley, CA 94720, United States of America
-    index: 1
-date: 21 April 2024
-bibliography: paper.bib
----
-
-# Summary
-
-T and B lymphocytes exhibit diverse surface receptor repertoires that interact with antigens to communicate and protect the body from foreign invaders. [@den2014activation] Modern methods allows the experimental characterization of the collective identity of the immune cells in tissue samples by sequencing RNA transcripts and cell receptors with scRNA-seq and scTCR-seq / scBCR-seq. Such methods generate high-dimensional data which are analyzed with the aid of dimensional reduction techniques to allow annotation of cell subsets. [@huang2022role]
-
-The R [@r2023r] package APackOfTheClones implements and *extends* a novel method to visualize clonal expansion at a single-cell resolution using circle packing, along with many clonal analysis utilities. Clonotype frequencies for each cell subset is counted and the values are used as radii and packed into one circular cluster with the largest circles near the center, and overlaid onto a the centroids of each cell subset on the corresponding 2D dimensional reduction plot. The package easily integrates into existing analysis pipelines using the *Seurat* [@hao2023dictionary] and *scRepertoire* [@borcherding2020screpertoire] packages. The original implementation was created in the julia language [@benzanson2017julia] by Christian Murray and Ben Murrell and successfully utilized in two immunology papers. [@ma2021single; @stark2022recombinant]
-
-<!-- should those last two sentences be here or in statement of need? -->
-
-# Statement of need
-
-Cells in single-cell experiments are conventionally visualized on a reduction scatterplot where each point represents a cell after all its features have been projected into two dimensions. The points can be colored by different factors to display useful information, including its assigned identity. [@andrews2021tutorial]
-
-Understanding the role of specific cells in various contexts requires understanding of relationships between cell populations and their behaviors. One attribute is the clonal expansion dynamics, which are inferred from downstream analyses. [@adams2020clonal] Overlaying clonal expansion information on a by-clonotype basis onto the reduction scatterplot of the cells of an experiment aids in adding an additional layer of insight, allowing for a swift, intuitive understanding of how clonal dynamics relate to the identified cellular subsets. For example, it can help gauge the presence of hyper expanded clones for each cell type; compare potential changes in frequencies after certain therapeutic treatments, etc.
-
-There is no standardized convention to visualize this single cell level clonotype data on the dimensional reduction. Some of the current approaches include 1. Using a color gradient corresponding to each frequency to highlight each individual point by the clonal expansion, implemented in *scRepertoire* and *scirpy* [sturm2020scirpy] 2. Overlaying a 2D contour where points representing clones with higher frequencies have elevated levels [@andreatta2023tcell] 3. Increasing sizes of points based on clonal frequency, used in figure 2c of @wang2021single.
-
[…]
