+++ dev/pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   APackOfTheClones-utils: APackOfTheClones-utils.html
   APackOfTheClones: APackOfTheClones.html
   v0-main: archive/v0-main.html
-last_built: 2024-04-22T05:38Z
+last_built: 2024-06-29T06:39Z
 urls:
   reference: https://qile0317.github.io/APackOfTheClones/reference
   article: https://qile0317.github.io/APackOfTheClones/articles
