+++ joss/paper.md
@@ -47,6 +47,34 @@ The main clonal expansion visualization that APackOfTheClones implements is show
 
 ![A single-cell experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
 
+<!-- ```R
+# the following code was used to generate the plot above - should it be included?
+
+# assume the example `contig_list` and `pbmc` (seurat object) are loaded
+
+contig_list %>%
+  combineTCR(
+    samples = c("P17B", "P17L", "P18B", "P18L", "P19B", "P19L", "P20B", "P20L"),
+    removeNA = FALSE,
+    removeMulti = FALSE,
+    filterMulti = FALSE
+  ) %>%
+  combineExpression(pbmc) %>%
+  RunAPOTC() %>%
+  AdjustAPOTC(
+    new_rad_scale_factor = 0.9,
+    nudge_cluster = c(8, 6, 2, 16, 12, 13, 10, 15, 14, 17, 11),
+    nudge_vector = list(c(-3,0), c(-1,2), c(-1.25,-1), c(0,-1), c(2, 0), c(1, 0), c(1.3, 0), c(1,0.5), c(2.6, 0), c(2, 0.2), c(3, 0.2))
+  ) %>%
+  APOTCPlot(
+    legend_position = "bottom left",
+    legend_sizes = c(1, 50, 200),
+    add_legend_background = FALSE,
+    use_default_theme = FALSE,
+    retain_axis_scales = TRUE
+  )
+``` -->
+
 The visualization gives the immediate insight that certain cell subsets such as those in cluster four contains many more expanded clones both by quantity and frequency.
 
 The package extends objects and the functionality of the *Seurat* and *scRepertoire* package, and given a correctly processed seurat object of scRNA-seq data that was combined with paired TCR/BCRs, only a few functions need to be used to as little or as much customization of function arguments as needed to produce a `ggplot` object [@wickham2016ggplot2] that fits into the conventional plotting ecosystem of R. Functions are accelerated with a `c++` layer via the *Rcpp* package [@eddelbuettel2011rcpp] to deliver all plots and R objects quickly in time complexity roughly linearly proportional to the number of cells, with the main time bottleneck being the plot display time.
