### 2024-07-24 #6868 @Qile0317
> @Qile0317 - can you please elaborate - why do you think this won't align with JOSS principles?

whoops - missed this comment completely. I think when I wrote this I was thinking about some comment on JOSS docs about how papers should be about the implementation of softwares. But I guess this sourcecode would be a part of the implementation. Although I think for now its probably fine.
