### 2024-08-05 #6868 @jspaezp
Sorry for the delay, Here are my current notes blocking the checking of all the fields (plus other suggestions, I think it is mostly there). Let me know if you believe any of them is not fair in any way and I would be happy to discuss further (This time with a much better turnaround time ...)

L20-22 [Nitpick] - I do not believe this statement should be part of the summary. I would encourage moving it to the statement of need (and paraphrase accordingly). [Revision: Same as the authors mention in the source code of the paper]
L39 [Please fix] - Unformatted reference.
42 [Suggestion] - I do not believe this sentence conveys the value provided by the package. I would appreciate a simple example (For example, depicting clones as contour lines does not allow the identification of xxx which would be clear in our viz because of yyy, as seen in Fig 1.)
62 [Suggestion] - I would love to see a performance comparison for that statement.

General:
I would appreciate a code snippet to generate the figure. [Revision: After reading the code referenced I believe you agree, I would also add the data loading to the snippet or reference to a documentation site on how to generate it, I believe this is the example in the landing page of the doc site]

> - should I write about other novelty features - customizing, highlighting, clone links?
> - maybe an "implementation details" section about how the user can store "runs" of this plot with different parameters and manually customize them?

I don’t believe that highlighting those features is strictly required (Although I would encourage a link to a documentation site with those examples)

Repository Notes:

The `LICENSE` file does not state an OSS license but the `LICENSE.md` does, please remove the former.
Other than that I would like to praise you, I believe the repository is excellent in terms of organization, testing and infrastructure. REALLY good work in this regard.
