### 2024-08-05 #6868 @Qile0317
Thanks for the reply @jspaezp Will get to solving each of your recommendations. Would you mind possibly splitting those up and creating github issues?

Regarding the L62 suggestion - what do you specifically mean by a "performance comparison"? Since there no alternative / competing programs to compare to aside from the original person's julia code which is not public. Do you mean a rough time vs input size (number of cells) graph? Theoretical time complexity?

And regarding the `LICENSE` file - I'm not 100% sure on this but I think it is some really weird R package convention to have a seperate `LICENSE.md` and `LICENSE` file and I see this everywhere - would you be okay with keeping it as is?
