### 2024-08-06 #6868 @jspaezp
@skanwal Thank you very much for that! I will go ahead and comment on those issues as needed.

@Qile0317 Regarding the license issue
> And regarding the LICENSE file - I'm not 100% sure on this but I think it is some really weird R package convention to have a seperate LICENSE.md and LICENSE file and I see this everywhere - would you be okay with keeping it as is?

You are totally right! It does seem to be a common pattern in R, please disregard.

Regarding the performance claim:
> Regarding the L62 suggestion - what do you specifically mean by a "performance comparison"? I believe it wasn't quite a "performance claim" that I made? Otherwise, since there no alternative / competing programs to compare to aside from the original person's julia code which is not public. Do you mean a rough time vs input size (number of cells) graph? Theoretical time complexity?

I agree that the mention of the high performance language is not a claim and thus does not 'require' a benchmark but highlights the effort and payoff of the process (again, this is not required, I was hoping you had this available to you)
