+++ joss/paper.md
@@ -37,7 +37,9 @@ Understanding the role of specific cells in various contexts requires understand
 
 There is no standardized convention to visualize this single cell level clonotype data on the dimensional reduction. Some of the current approaches include 1. Using a color gradient corresponding to each frequency to highlight each individual point by the clonal expansion, implemented in *scRepertoire* and *scirpy* [@sturm2020scirpy] 2. Overlaying a 2D contour where points representing clones with higher frequencies have elevated levels [@andreatta2023tcell] 3. Increasing sizes of points based on clonal frequency, used in figure 2c of @wang2021single.
 
-These approaches capture approximate, global trends but do not provide an exact representation. They lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. From a visual standpoint, This limitation hinders subtle variations in clonal expansion patterns. APackOfTheClones solves this issue by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses.
+These approaches capture approximate, global trends but do not provide an exact representation. They lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. From a visual standpoint, This limitation hinders subtle variations in clonal expansion patterns.
+
+APackOfTheClones offers an approach to address these issues by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses. The original ideation and implementation was done by Christian Murray and Ben Murrell in the julia language [@benzanson2017julia] and utilized successfully in two immunology papers which showcases their potential. [@ma2021single; @stark2022recombinant] However, language choice and the lack of a user-friendly interface limited its accessibility and integration into existing pipelines. The integration of this approach within Seurat and scRepertoire in the popular R ecosystem bridges this gap, increasing accessibility to the wider bioinformatics community through a straightforward, familiar interface.
 
 APackOfTheClones also offers a suite of methods for visualizing and analyzing single-cell clonal data. Novelty features include functions for highlighting certain clones, and the filtering and visualization of clonotypes shared between subsets by linking circles on the APackOfTheClones clonal expansion plot.
 
