+++ joss/paper.md
@@ -45,7 +45,7 @@ APackOfTheClones also offers a suite of methods for visualizing and analyzing si
 
 # Results
 
-The main clonal expansion visualization that APackOfTheClones implements is shown in the following figure \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
+The main clonal expansion visualization that APackOfTheClones implements is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
 
 ![A single-cell experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
 
