+++ joss/paper.md
@@ -39,13 +39,13 @@ There is no standardized convention to visualize this single cell level clonotyp
 
 These approaches capture approximate, global trends but do not provide an exact representation. They lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. From a visual standpoint, This limitation hinders subtle variations in clonal expansion patterns.
 
-APackOfTheClones offers an approach to address these issues by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses. The original ideation and implementation was done by Christian Murray and Ben Murrell in the julia language [@benzanson2017julia] and utilized successfully in two immunology papers which showcases their potential. [@ma2021single; @stark2022recombinant] However, language choice and the lack of a user-friendly interface limited its accessibility and integration into existing pipelines. The integration of this approach within Seurat and scRepertoire in the popular R ecosystem bridges this gap, increasing accessibility to the wider bioinformatics community through a straightforward, familiar interface.
+APackOfTheClones offers an approach to address these issues by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses. The original julia implementation's successful utilization in two publication is promising evidence in its potential [@ma2021single; @stark2022recombinant] - however, language choice and the lack of a user-friendly interface limited its accessibility and integration into existing pipelines. The integration of this approach within Seurat and scRepertoire in the popular R ecosystem bridges this gap, increasing accessibility to the wider bioinformatics community through a straightforward, familiar interface.
 
 APackOfTheClones also offers a suite of methods for visualizing and analyzing single-cell clonal data. Novelty features include functions for highlighting certain clones, and the filtering and visualization of clonotypes shared between subsets by linking circles on the APackOfTheClones clonal expansion plot.
 
 # Results
 
-The main clonal expansion visualization that APackOfTheClones implements is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
+An example of the main clonal expansion visualization that APackOfTheClones can produce is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
 
 ![A single-cell experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
 
