### 2024-08-21 f384e1d
added abbreviations
