### 2024-08-21 e31503a
attempted improvement on L42
