+++ joss/paper.md
@@ -37,7 +37,8 @@ Understanding the role of specific cells in various contexts requires understand
 
 There is no standardized convention to visualize this on the dimensional reduction. Some of the current approaches include 1. Using a color gradient corresponding to each frequency to highlight each individual point by the CE, implemented in *scRepertoire* and *scirpy* [@sturm2020scirpy] 2. Overlaying a 2D contour where points representing clones with higher frequencies have elevated levels [@andreatta2023tcell] 3. Increasing sizes of points based on clonal frequency, used in figure 2c of @wang2021single.
 
-These approaches capture approximate, global trends but do not provide an exact representation. They lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. From a visual standpoint, This limitation hinders subtle variations in CE patterns.
+From a visual standpoint, these approaches capture approximate trends that lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. For instance, using a color gradient or contour lines to depict clones fails to directly differentiate between different clonotypes within the same cell subset. This is particularly problematic in scenarios where identifying rare clonotypes is essential, such as detecting the presence of emerging clones in response to a therapy or tracking the evolution of immune responses in different contexts.
+<!-- pretty unsure about this paragraph and whether it captures enough:P -->
 
 APackOfTheClones offers an approach to address these issues by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses. The original julia implementation's successful utilization in two publication is promising evidence in its potential [@ma2021single; @stark2022recombinant] - however, language choice and the lack of a user-friendly interface limited its accessibility and integration into existing pipelines. The integration of this approach within Seurat and scRepertoire in the popular R ecosystem bridges this gap, increasing accessibility to the wider bioinformatics community through a straightforward, familiar interface.
 
