### 2024-08-21 #6868 @Qile0317
@jspaezp I have made fixes to some of the mentions - let me know what you think. Regarding the L42 suggestion - please let me know in the github issue what you think of my additions. I additionally thought there was maybe a bit too much repetition so I added abbreviations for single-cell as SC and clonal expansion as CE. Lastly, I contaced @ncborcherding to potentially upload the example object to zenodo or some other archiving repo so the data can be referenced in the example code. Let me know if there is anything else i can help with / improve. Thanks.
