### 2024-09-06 #6868 @Qile0317
@jspaezp any updates on your thoughts? The following is a direct reference to the L42 change: <https://github.com/Qile0317/APackOfTheClones/commit/e31503a15b571f424cba09fb84d468e1eaa08cb6#diff-407d3babcb08d5dd50a996c55e76fdc15d54ca21cf0e8dc50cd0fc7ef505315aR40>

Looks like the data can't be archived for a variety of reasons but I will write in more detail how to get it to work.
