### 2024-09-06 #6868 @jspaezp
Yeah I thought that was good, clarifying what aspects in traditional viz could confound what you are trying tom measure was good! Any updates on the dataset to make the visualization end-to-end reproducible?
