+++ joss/paper.md
@@ -46,39 +46,9 @@ APackOfTheClones also offers a suite of methods for visualizing and analyzing SC
 
 # Results
 
-An example of the main CE visualization that APackOfTheClones can produce is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping.
-
-<!-- The source code to generate this plot is ... -->
+An example of the main CE visualization that APackOfTheClones can produce is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping. The steps needed to generate the plot can be found in the [supplemental repository](https://github.com/Qile0317/APackOfTheClonesSupplemental)
 
 ![A scRNA-seq experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
-
-<!-- ```R
-# the following code was used to generate the plot above - should it be included?
-
-# assume the example `contig_list` and `pbmc` (seurat object) are loaded
-
-contig_list %>%
-  combineTCR(
-    samples = c("P17B", "P17L", "P18B", "P18L", "P19B", "P19L", "P20B", "P20L"),
-    removeNA = FALSE,
-    removeMulti = FALSE,
-    filterMulti = FALSE
-  ) %>%
-  combineExpression(pbmc) %>%
-  RunAPOTC() %>%
-  AdjustAPOTC(
-    new_rad_scale_factor = 0.9,
-    nudge_cluster = c(8, 6, 2, 16, 12, 13, 10, 15, 14, 17, 11),
-    nudge_vector = list(c(-3,0), c(-1,2), c(-1.25,-1), c(0,-1), c(2, 0), c(1, 0), c(1.3, 0), c(1,0.5), c(2.6, 0), c(2, 0.2), c(3, 0.2))
-  ) %>%
-  APOTCPlot(
-    legend_position = "bottom left",
-    legend_sizes = c(1, 50, 200),
-    add_legend_background = FALSE,
-    use_default_theme = FALSE,
-    retain_axis_scales = TRUE
-  )
-``` -->
 
 The visualization gives the immediate insight that certain cell subsets such as those in cluster four contains many more expanded clones both by quantity and frequency.
 
@@ -96,5 +66,7 @@ APackOfTheClones offers a fast, and simple interface to produce an intuitive, ea
 # Acknowledgements
 
 Thanks to Ben Murrell for introducing the idea, implementing the original julia code along with Christian Murray, as well as giving debug support and suggestions. Thanks to Nick Borcherding for providing more insights, suggestions, and promotion.
+
+# Supplemental Materials
 
 # References
