+++ joss/paper.md
@@ -46,7 +46,7 @@ APackOfTheClones also offers a suite of methods for visualizing and analyzing SC
 
 # Results
 
-An example of the main CE visualization that APackOfTheClones can produce is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping. The steps needed to generate the plot can be found in the [supplemental repository](https://github.com/Qile0317/APackOfTheClonesSupplemental)
+An example of the main CE visualization that APackOfTheClones can produce is shown in the following \autoref{fig:example}, using scRNA-seq + scTCR-seq data from @borcherding2021mapping. The steps needed to generate the plot can be found in the [supplemental repository](https://github.com/Qile0317/APackOfTheClonesAppendix)
 
 ![A scRNA-seq experiment's dimensional reduction projection and its corresponding APackOfTheClones clonal expansion plot. The plot on the right is the projection of all cells of its *Seurat* object based on Uniform Manifold Approximation and Projection (UMAP), colored by unknown identities 1 through 17. On the left is the clonal expansion plot generated by APackOfTheClones on the same data. Each cell identity on the seurat object corresponds to a cluster of circles with a similar geometric placement, and the size of each individual circle is the clonotype frequency of some clonotype within that cell subset. Note that the largest clones are packed near the origins of each cluster to accentuate their difference with the rest of all clonotypes. \label{fig:example}](figures/main_example.png)
 
@@ -69,6 +69,6 @@ Thanks to Ben Murrell for introducing the idea, implementing the original julia 
 
 # Supplemental Materials
 
-[Supplemental information repository](https://github.com/Qile0317/APackOfTheClonesSupplemental)
+[Supplemental information repository](https://github.com/Qile0317/APackOfTheClonesAppendix)
 
 # References
