### 2024-09-14 dea522f
made links explicit in joss paper
