### 2024-09-14 #6868 @Qile0317
@skanwal I believe the next step is to archive the repository on something like zenodo? If I am allowed manually make some changes to the code before doing so let me know. There are things like unnessecary comments and hidden files, etc. I plan to remove. Also - does the joss folder need to be in there when archiving?
