### 2024-10-03 #6868 @Qile0317
@skanwal Would it be okay if the zenodo archive was not a carbon copy of the current repository and I cleaned up and modified some files?
