+++ R/RunAPOTC.R
@@ -9,7 +9,7 @@
 #' clonal expansion plot ([APOTCPlot]) and stores it in the seurat object.
 #' Gets sizes of unique clones and utilizes a circle-packing algorithm to
 #' pack circles representing individual clones in approximately the same
-#' dimensional reduction (`reduction_base`) coordinates based on some cell 
+#' dimensional reduction (`reduction_base`) coordinates based on some cell
 #' ident (defaults to the active ident).
 #'
 #' The parameter `extra_filter` along with an unlimited number of additional
@@ -167,7 +167,7 @@ RunAPOTC <- function(
     clonecall = "strict",
     ...,
     extra_filter = NULL,
-    alt_ident = NULL, # TODO check
+    alt_ident = NULL,
     run_id = NULL,
 
     clone_scale_factor = "auto",
@@ -192,8 +192,8 @@ RunAPOTC <- function(
     # compute/check inputs
     reduction_base <- attempt_correction(seurat_obj, reduction_base)
     clonecall <- .theCall(seurat_obj@meta.data, clonecall)
-    # TODO check alt_ident
-    
+
+   
     if (should_estimate(clone_scale_factor)) {
         clone_scale_factor <- estimate_clone_scale_factor(seurat_obj, clonecall)
         if (verbose) message(paste(
@@ -250,7 +250,7 @@ RunAPOTC <- function(
     # store the apotc object in the correct slot with the correct id
     seurat_obj <- setApotcData(seurat_obj, obj_id, apotc_obj)
 
-    # TODO this could be simplified, something with looking at the grandparent
+
     # environment frame with n = 2
     command_obj <- make_apotc_command(call_time)
     seurat_obj <- log_seurat_command(
@@ -265,7 +265,7 @@ RunAPOTC <- function(
 
 # # a super simple regression of manually determined clone scale based on cell
 # # count a much more complicated model can use other facts about the seurat
-# # obj to improve how visually plesant is it. Some overlap factor could 
+# # obj to improve how visually plesant is it. Some overlap factor could
 # # probably be eestimated with the raw clone counts
 # cell_count <- c(80, 365, 2500)
 # desirable_factor <- c(1, 0.3, 0.2)
@@ -312,7 +312,7 @@ check_apotc_identifiers <- function(args) {
         stop(call. = FALSE, "`seurat_obj` must be a Seurat object.")
     }
 
-    # TODO nulls shouldnt be allowed in some cases
+
     typecheck(args$reduction_base, is_a_character, is.null)
     typecheck(args$clonecall, is_a_character, is.null)
     typecheck(args$extra_filter, is_character, is.null)
@@ -357,5 +357,5 @@ RunAPOTC_parameter_checker <- function() {
         ))
     }
 
-	# TODO more checks
+
 }
