### 2024-10-03 #6868 @Qile0317
Just for my own reference, here is my personal post-review checklist:
- [  ] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
