### 2024-10-09 #6868 @skanwal
> Or should I just instead make it a new version like 1.2.1?

Think that'll be appropriate for the changes you've linked. Thanks.
