+++ tests/testthat/test-Repulsion.R
@@ -58,10 +58,10 @@ test_that("calculate_repulsion_vectors() works", {
     calculate_repulsion_vectors(
       overall_repulsion_vec, inp, 3
     ),
-    list(list(c(0, 0), c(-0.069912727392497, -0.0873909092406212), 
-              c(-0.100898062586345, -0.134530750115127)), list(c(0.0699127273924969, 
+    list(list(c(0, 0), c(-0.069912727392497, -0.0873909092406212),
+              c(-0.100898062586345, -0.134530750115127)), list(c(0.0699127273924969,
                                                                  0.0873909092406212), c(0, 0), c(1.48636883882061, 1.48636883882061
-                                                                 )), list(c(0.100898062586345, 0.134530750115127), c(-1.48636883882061, 
+                                                                 )), list(c(0.100898062586345, 0.134530750115127), c(-1.48636883882061,
                                                                                                                      -1.48636883882061), c(0, 0))),
     tolerance = 1e-6
   )
@@ -85,7 +85,7 @@ test_that("calculate_transformation_vectors() works", {
     ),
     tolerance = 1e-6
   )
-  
+ 
   expect_identical(
     calculate_transformation_vectors(
       initialize_direction_vectors(2),
@@ -106,19 +106,19 @@ test_that("repulse_cluster() works", {
       thr = 1, G = 1, max_iter = 10,
       verbose = FALSE
     ),
-    list(list(x = c(-1.85085353933408, 0.749146460665919, -0.550853539334081, 
-                    -2.52812593801653, -1.27911424816896, 0.698158150513489, 1.99266321363309, 
-                    -0.550853539334081, 3.03766022320323), y = c(-0.0561113921339256, 
-                                                                 -0.0561113921339256, -1.95347798823495, -2.25413370730339, -3.81617397522996, 
+    list(list(x = c(-1.85085353933408, 0.749146460665919, -0.550853539334081,
+                    -2.52812593801653, -1.27911424816896, 0.698158150513489, 1.99266321363309,
+                    -0.550853539334081, 3.03766022320323), y = c(-0.0561113921339256,
+                                                                 -0.0561113921339256, -1.95347798823495, -2.25413370730339, -3.81617397522996,
                                                                  -3.51551825616153, -1.99096696428653, 1.8412552039671, -0.285686562126976
-                    ), rad = c(1.3, 1.3, 1, 1, 1, 1, 1, 1, 1), centroid = c(-0.550853539334081, 
-                                                                            -0.688566924167602), clRad = 4.58851376253731), list(x = c(3.25085353933408, 
-                                                                                                                                       5.85085353933408, 4.55085353933408, 2.57358114065163, 3.8225928304992, 
+                    ), rad = c(1.3, 1.3, 1, 1, 1, 1, 1, 1, 1), centroid = c(-0.550853539334081,
+                                                                            -0.688566924167602), clRad = 4.58851376253731), list(x = c(3.25085353933408,
+                                                                                                                                       5.85085353933408, 4.55085353933408, 2.57358114065163, 3.8225928304992,
                                                                                                                                        5.79986522918165, 7.09437029230125, 4.55085353933408, 8.13936730187139
-                                                                            ), y = c(6.32102245620128, 6.32102245620128, 4.42365586010025, 
-                                                                                     4.12300014103181, 2.56095987310524, 2.86161559217367, 4.38616688404867, 
-                                                                                     8.2183890523023, 6.09144728620823), rad = c(1.3, 1.3, 1, 1, 1, 
-                                                                                                                                 1, 1, 1, 1), centroid = c(4.55085353933408, 5.6885669241676), 
+                                                                            ), y = c(6.32102245620128, 6.32102245620128, 4.42365586010025,
[…]
