### 2024-10-09 f1fd669
rm some TODOs and FIXMEs
