+++ NAMESPACE
@@ -20,5 +20,6 @@ export(renameApotcRun)
 export(showCloneHighlight)
 export(vizAPOTC)
 importFrom(Rcpp,sourceCpp)
+importFrom(assertthat,assert_that)
 importFrom(dplyr,"%>%")
 useDynLib(APackOfTheClones, .registration = TRUE)
