### 2024-10-09 2c31b65
update citation, incr patch in NEWs but not DESCRIPTION
