+++ NEWS.md
@@ -1,11 +1,12 @@
-# APackOfTheClones 1.2.0.9000 (development version)
+# APackOfTheClones 1.2.1
 
 ## Additions
 
-<!-- - A JOSS paper for this package has now been published - the corresponding README badge and citation information (`citation("APackOfTheClones")`) have been added. -->
+- Add code of conduct and contributing guidelines in the package source code.
 
 ## Changes
 
+- updated citation information upon calling `citation("APackOfTheClones")`: the package as of this patch version is undergoing review in the [journal of open source software](https://joss.theoj.org/papers/eb9d2b2fc3ffb1d440bd7120b66f1f71).
 - Added a check within an internal function to ensure that certain columns are not present in the metadata of the seurat object before proceeding with the clonal expansion plot.
 - Spelling fixes to function documentation and vignettes.
 - Removed package startup deprecation message for returning users of version `0.1.x`.
