### 2024-10-09 82c97e2
remove redundant imports, update citation with CRAN DOI
