+++ DESCRIPTION
@@ -34,10 +34,8 @@ Imports:
     dplyr,
     ggforce,
     ggplot2,
-    grDevices,
     hash,
     lifecycle,
-    methods,
     Rcpp,
     rlang,
     Seurat,
