### 2024-10-09 c52f764
modified R CMD check action to only check oldrel1
