+++ .github/workflows/R-CMD-check.yaml
@@ -1,8 +1,8 @@
 on:
   push:
-    branches: [main, dev, unstable]
+    branches: [main, dev]
   pull_request:
-    branches: [main, dev, unstable]
+    branches: [main, dev]
 
 name: R-CMD-check
 
@@ -22,9 +22,10 @@ jobs:
           - {os: windows-latest, r: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
          
-          # check compatibility with older releases
+          # check compatibility with oldrel-1
+          - {os: macos-latest,   r: 'oldrel-1'}
+          - {os: windows-latest, r: 'oldrel-1'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
-          - {os: ubuntu-latest,   r: 'oldrel-2'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
