### 2024-10-09 ff6877c
redid packing algorithm to take into account double precision, updated test data. 3 cases fail
