+++ tests/testthat/test-showCloneHighlight.R
@@ -1,8 +1,5 @@
-test_that("vdiffr test of showCloneHighlight works", {
-
-    skip_on_ci() # there is no visually distinguishable difference in the artifact
-
-    test_cl_5_9_highlighted_ggplot <- get(data("combined_pbmc")) %>%
+runTestShowCloneHighlight <- function() {
+    get(data("combined_pbmc")) %>%
         vizAPOTC(
             clonecall = "aa",
             seurat_clusters = c(5, 9),
@@ -13,36 +10,50 @@ test_that("vdiffr test of showCloneHighlight works", {
         showCloneHighlight(c(
             "CASLSGSARQLTF_CASSPTVAGEQFF", "CAELNQAGTALIF_CASSQAPFSTSGELFF"
         ))
+}
+
+test_that("vdiffr test of showCloneHighlight works", {
+
+    #skip_on_ci() # there is no visually distinguishable difference in the artifact
 
     expect_doppelganger(
         "showCloneHighlight_only_clusters_5_9",
-        test_cl_5_9_highlighted_ggplot
+        runTestShowCloneHighlight()
     )
 
 })
 
 test_that("showCloneHighlight works", {
 
-    data("combined_pbmc")
+    test_cl_5_9_highlighted_data <-
+        get_ggplot_data(runTestShowCloneHighlight())
 
-    test_cl_5_9_highlighted_data <- combined_pbmc %>%
-        vizAPOTC(
-            clonecall = "aa",
-            seurat_clusters = c(5, 9),
-            retain_axis_scales = TRUE,
-            add_size_legend = FALSE,
-            verbose = FALSE
-        ) %>%
-        showCloneHighlight(c(
-            "CASLSGSARQLTF_CASSPTVAGEQFF", "CAELNQAGTALIF_CASSQAPFSTSGELFF"
-        )) %>%
-        get_ggplot_data()
-
+    # check the cluster, clonotype (and its color), and corresponding radii
+    # are setequal since order doesn't matter
     expect_equal(
-        object = test_cl_5_9_highlighted_data,
+        object = test_cl_5_9_highlighted_data %>%
+            dplyr::select(label, color, clonotype, r) %>%
+            dplyr::arrange(clonotype),
         expected = getdata(
             "showCloneHighlight", "cl_5_9_highlighted_data"
-        )
+        ) %>%
+            dplyr::select(label, color, clonotype, r) %>%
+            dplyr::arrange(clonotype)
+    )
+
+    # check that for each cluster and radii sets, there are setequal (x, y)
+    # pairs since it would be geometrically equal
+    expect_equal(
+        object = test_cl_5_9_highlighted_data %>%
+            dplyr::select(label, x, y, r) %>%
+            dplyr::arrange(x),
+        expected =
+            getdata(
+                "showCloneHighlight", "cl_5_9_highlighted_data"
+            ) %>%
+            dplyr::select(label, x, y, r) %>%
+            dplyr::arrange(x),
+        tolerance = 1e-5
     )
 
     # TODO test S4 also present in scRepertoire
