+++ R/clusters.R
@@ -32,6 +32,12 @@ set_cluster_radius <- function(l, v) {l$clRad <- v; l}
 set_clonotypes <- function(l, v) {l$clonotype <- v; l}
 
 # convert clusterlist to dataframe, assuming its ***valid***
+# returns a dataframe with columns:
+# - label
+# - x
+# - y
+# - r
+# - clonotype
 convert_to_dataframe <- function(
     clstr_list, seurat_cluster_index, detail = TRUE
 ) {
