### 2024-10-10 ee61ebb
more verbosity options for helper
