+++ R/ApotcData.R
@@ -429,3 +429,18 @@ set_labels <- function(apotc_obj, x) {
 	apotc_obj@labels <- x
 	apotc_obj
 }
+
+# functions for testing
+
+does_apotc_data_clusterlist_equal_expected <- function(
+    apotc_obj, expected_clusterlist, tolerance = 1e-6, verbose = FALSE
+) {
+    areGeometricallyEqualClusterLists(
+        a = get_clusterlists(apotc_obj),
+        b = expected_clusterlist,
+        rad_decrease = get_rad_decrease(apotc_obj),
+        clone_scale_factor = get_clone_scale_factor(apotc_obj),
+        tolerance = tolerance,
+        verbose = verbose
+    )
+}
