+++ R/APackOfTheClones-package.R
@@ -4,6 +4,7 @@
 ## usethis namespace: start
 #' @importFrom assertthat assert_that
 #' @importFrom dplyr %>%
+#' @importFrom glue glue
 #' @importFrom Rcpp sourceCpp
 #' @useDynLib APackOfTheClones, .registration = TRUE
 ## usethis namespace: end
