+++ NAMESPACE
@@ -22,4 +22,5 @@ export(vizAPOTC)
 importFrom(Rcpp,sourceCpp)
 importFrom(assertthat,assert_that)
 importFrom(dplyr,"%>%")
+importFrom(glue,glue)
 useDynLib(APackOfTheClones, .registration = TRUE)
