### 2024-10-10 4be31ec
improved and more fixes for cluster lsit checker
