+++ R/clusters.R
@@ -9,14 +9,19 @@
 # [["clonotype"]] clonotype based on original clonecall
 
 # getters for a single clusterlist
+# does not handle the case of empty clusterlists
 
-get_x_coords <- function(l) l[[1]]
-get_y_coords <- function(l) l[[2]]
-get_radii <- function(l) l[[3]]
-get_centroid <- function(l) l[[4]]
+makeGetter <- function(key) {
+    function(obj) if (isnt_empty(obj)) obj[[key]]
+}
+
+get_x_coords <- makeGetter(1)
+get_y_coords <- makeGetter(2)
+get_radii <- makeGetter(3)
+get_centroid <- makeGetter(4)
 get_centroid_x <- function(l) get_centroid(l)[1]
 get_centroid_y <- function(l) get_centroid(l)[2]
-get_cluster_radius <- function(l) l[[5]]
+get_cluster_radius <- makeGetter(5)
 get_clonotypes <- function(l) l$clonotype # not index due to legacy clusterlists
 
 get_num_clones <- function(l) length(get_x_coords(l))
@@ -24,12 +29,20 @@ get_num_clones <- function(l) length(get_x_coords(l))
 contains_clonotypes <- function(x) !is.null(get_clonotypes(x))
 
 # setters for a single clusterlist
-set_x_coords <- function(l, v) {l$x <- v; l}
-set_y_coords <- function(l, v) {l$y <- v; l}
-set_radii <- function(l, v) {l$rad <- v; l}
-set_centroid <- function(l, v) {l$centroid <- v; l}
-set_cluster_radius <- function(l, v) {l$clRad <- v; l}
-set_clonotypes <- function(l, v) {l$clonotype <- v; l}
+
+makeSetter <- function(key) {
+    function(obj, value) {
+        if (isnt_empty(obj)) obj[[key]] <- value
+        obj
+    }
+}
+
+set_x_coords <- makeSetter(1)
+set_y_coords <- makeSetter(2)
+set_radii <- makeSetter(3)
+set_centroid <- makeSetter(4)
+set_cluster_radius <- makeSetter(5)
+set_clonotypes <- makeSetter("clonotype")
 
 # convert clusterlist to dataframe, assuming its ***valid***
 # returns a dataframe with columns:
@@ -150,14 +163,104 @@ read_centroids <- function(list_of_clusterlists) {
   })
 }
 
+# clusterlist checkers below
+
+isValidClusterList <- function(x, legacy = FALSE) {
+
+    if (identical(x, list())) return(TRUE)
+
+    basicChecks <- validate_that(
+        is.list(x),
+        (length(x) >= (6 - legacy)),
+        is.numeric(x[[1]]),
+        is.numeric(x[[2]]),
+        is.numeric(x[[3]]),
+        is.numeric(x[[4]]),
+        is.numeric(x[[5]]),
+        (legacy || is.character(x[[6]]))
+    )
+
+    if (!isTRUE(basicChecks)) {message(basicChecks); return(FALSE)}
+
+    lengthChecks <- validate_that(
+        length(unique(sapply(x[c(1, 2, 3, if (legacy) NULL else 6)], length))) == 1,
+        length(x[[4]]) == 2,
+        length(x[[5]]) == 1
+    )
+    
+    if (!isTRUE(lengthChecks)) {message(lengthChecks); return(FALSE)}
+
+    nameChecks <- identical(
+        names(x)[1:5],
+        c("x", "y", "rad", "centroid", "clRad")
+    )
+    if (!legacy) {
+        nameChecks <- nameChecks && identical(names(x)[6], "clonotype")
+    }
+
+    if (isFALSE(nameChecks)) {
[…]
