+++ NAMESPACE
@@ -21,6 +21,7 @@ export(showCloneHighlight)
 export(vizAPOTC)
 importFrom(Rcpp,sourceCpp)
 importFrom(assertthat,assert_that)
+importFrom(assertthat,validate_that)
 importFrom(dplyr,"%>%")
 importFrom(glue,glue)
 useDynLib(APackOfTheClones, .registration = TRUE)

[…]
