### 2024-10-10 dd46f5d
delete all additional validation
