+++ R/clusters.R
@@ -210,189 +210,7 @@ isValidListOfClusterLists <- function(x, legacy = FALSE, verbose = FALSE) {
         if (verbose) {
             message("clusterlist ", i, " is invalid")
         }
-        #print(summary(x[[i]]))
         return(FALSE)
     }
     return(TRUE)
-}
-
-areGeometricallyEqualListsOfClusterLists <- function(
-    a, b, rad_decrease, clone_scale_factor,
-    legacy = TRUE, tolerance = 1e-6, verbose = FALSE
-) {
-
-    if (!isValidListOfClusterLists(a, legacy, verbose)) {
-        if (verbose) message("first list of clusterlists is invalid")
-        return(FALSE)
-    }
-
-    if (!isValidListOfClusterLists(b, legacy, verbose)) {
-        if (verbose) message("second list of clusterlists is invalid")
-        return(FALSE)
-    }
-
-    if (length(a) != length(b)) {
-        if (verbose) message("length of clusterlists unequal")
-        return(FALSE)
-    }
-    
-    for (i in seq_along(a)) {
-        if (!areGeometricallyEqualClusterLists(
-            a[[i]], b[[i]], rad_decrease, clone_scale_factor, tolerance, verbose
-        )) {
-            if (verbose) {
-                message("clusterlists at ", i, " is not equal")
-            }
-            return(FALSE)
-        }
-    }
-
-    return(TRUE)
-}
-
-areGeometricallyEqualClusterLists <- function(
-    a, b, rad_decrease, clone_scale_factor,
-    legacy = TRUE, tolerance = 1e-6, verbose = FALSE
-) {
-
-    logIfVerbose <- function(...) if (verbose) message(glue(...))
-    logDiffIfVerbose <- function(...) {
-        logIfVerbose(getDiffAsListInStr(...))
-    }
-
-    if (is_empty(a) && is_empty(b)) return(TRUE)
-
-    if (get_num_clones(a) != get_num_clones(b)) {
-        logIfVerbose(
-            "number of clones unequal: ",
-            "{get_num_clones(a)}, {get_num_clones(b)}"
-        )
-        return(FALSE)
-    }
-
-
-    if (!haveSameElements(get_radii(a), get_radii(b))) {
-        logIfVerbose("radii don't contain the same elements")
-        logDiffIfVerbose(
-            get_radii(a), get_radii(b),
-            deparse(substitute(a)), deparse(substitute(b))
-        )
-        return(FALSE)
-    }
-
-    if (!legacy && !haveSameElements(get_clonotypes(a), get_clonotypes(b))) {
-        logDiffIfVerbose(
-            get_clonotypes(a), get_clonotypes(b),
-            deparse(substitute(a)), deparse(substitute(b))
-        )
-        return(FALSE)
-    }
-
-    list_of_normed_ab <- list(a, b) %>%
-        lapply(function(x)
-            normalizeClusterListRadii(x, rad_decrease, clone_scale_factor)
-        )
-    
-    list_of_normed_ab %>%
-        append(list(
-            "legacy" = legacy, "threshold" = tolerance, "verbose" = verbose
-        )) %>%
-        applyListAsArgsTo(areGeometricallyEqualNormalizedClusterLists)
-
-}
-
-normalizeClusterListRadii <- function(
-    clusterlist, rad_decrease, clone_scale_factor
-) {
-    clusterlist %>% set_radii(
-        ((get_radii(clusterlist) + rad_decrease) / clone_scale_factor)^2
-    )
-}
-
-# compare two clusterlists that have their radii transformed back into their
-# original clone counts.
-#
-# this assumes that the number of clones are equal, the radii are setequal,
-# and that the overall clonotypes are setequal.
-#
[…]
