### 2024-10-10 401c976
rm glue, update NEWS
