+++ R/APackOfTheClones-package.R
@@ -4,7 +4,6 @@
 ## usethis namespace: start
 #' @importFrom assertthat assert_that validate_that
 #' @importFrom dplyr %>%
-#' @importFrom glue glue
 #' @importFrom Rcpp sourceCpp
 #' @useDynLib APackOfTheClones, .registration = TRUE
 ## usethis namespace: end
