+++ NAMESPACE
@@ -23,5 +23,4 @@ importFrom(Rcpp,sourceCpp)
 importFrom(assertthat,assert_that)
 importFrom(assertthat,validate_that)
 importFrom(dplyr,"%>%")
-importFrom(glue,glue)
 useDynLib(APackOfTheClones, .registration = TRUE)
