+++ DESCRIPTION
@@ -34,7 +34,6 @@ Imports:
     dplyr,
     ggforce,
     ggplot2,
-    glue,
     hash,
     lifecycle,
     methods,
