### 2024-10-10 64ba8c9
slight fixes and test skips on macos
