+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: APackOfTheClones
 Type: Package
 Title: Visualization of Clonal Expansion for Single Cell Immune Profiles
-Version: 1.2.0.9000
+Version: 1.2.1
 Authors@R: c(
     person(
         "Qile",
