+++ dev/pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   APackOfTheClones-utils: APackOfTheClones-utils.html
   APackOfTheClones: APackOfTheClones.html
   archive/v0-main: archive/v0-main.html
-last_built: 2024-10-09T03:26Z
+last_built: 2024-10-10T21:55Z
 urls:
   reference: https://qile0317.github.io/APackOfTheClones/reference
   article: https://qile0317.github.io/APackOfTheClones/articles
