+++ CONTRIBUTING.md
@@ -5,7 +5,7 @@ The goal of this guide is to help you get up and contributing to APackOfTheClone
 1. Filing a bug report in an issue.
 2. Suggesting a change via a pull request.
 
-Please note that ggplot2 is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By contributing to this project,
+Please note that APackOfTheClones is released with a [Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By contributing to this project,
 you agree to abide by its terms.
 
 ## Bug Reports
@@ -69,7 +69,7 @@ Pull requests will be evaluated against a seven point checklist:
     need for change.
 
     Also include this motivation in `NEWS` so that when a new release of
-    ggplot2 comes out it's easy for users to see what's changed. Add your
+    APackOfTheClones comes out it's easy for users to see what's changed. Add your
     item at the top of the file and use markdown for formatting. The
     news item should end with `(@yourGithubUsername, #the_issue_number)`.
 
