+++ pkgdown.yml
@@ -2,10 +2,13 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
+  APackOfTheClones-install: APackOfTheClones-install.html
   APackOfTheClones-runs: APackOfTheClones-runs.html
+  APackOfTheClones-shared: APackOfTheClones-shared.html
+  APackOfTheClones-utils: APackOfTheClones-utils.html
   APackOfTheClones: APackOfTheClones.html
   archive/v0-main: archive/v0-main.html
-last_built: 2024-10-09T03:18Z
+last_built: 2024-10-10T22:18Z
 urls:
   reference: https://qile0317.github.io/APackOfTheClones/reference
   article: https://qile0317.github.io/APackOfTheClones/articles
