### 2024-11-06 3ce08de
Remove duplicate bib entries
Signed-off-by: Qile <43093685+Qile0317@users.noreply.github.com>
