+++ joss/paper.bib
@@ -1,15 +1,3 @@
-@article{den2014activation,
-  title={The activation of the adaptive immune system: cross-talk between antigen-presenting cells, T cells and B cells},
-  author={den Haan, Joke MM and Arens, Ramon and van Zelm, Menno C},
-  journal={Immunology letters},
-  volume={162},
-  number={2},
-  pages={103--112},
-  year={2014},
-  publisher={Elsevier},
-  doi={10.1016/j.imlet.2014.10.011}
-}
-
 @article{huang2022role,
   title={The role of single-cell profiling and deep immunophenotyping in understanding immune therapy cardiotoxicity},
   author={Huang, Yuhsin Vivian and Waliany, Sarah and Lee, Daniel and Galdos, Francisco X and Witteles, Ronald M and Neal, Joel W and Fan, Alice C and Maecker, Holden T and Nguyen, Patricia K and Wu, Sean M and others},
@@ -78,18 +66,6 @@
   year={2020},
   publisher={Oxford University Press},
   doi={10.37473/dac/10.1101/2020.04.10.035865}
-}
-
-@article{huang2022role,
-  title={The role of single-cell profiling and deep immunophenotyping in understanding immune therapy cardiotoxicity},
-  author={Huang, Yuhsin Vivian and Waliany, Sarah and Lee, Daniel and Galdos, Francisco X and Witteles, Ronald M and Neal, Joel W and Fan, Alice C and Maecker, Holden T and Nguyen, Patricia K and Wu, Sean M and others},
-  journal={Cardio Oncology},
-  volume={4},
-  number={5},
-  pages={629--634},
-  year={2022},
-  publisher={American College of Cardiology Foundation Washington DC},
-  doi={10.1016/j.jaccao.2022.08.012}
 }
 
 @article{andrews2021tutorial,
@@ -175,14 +151,6 @@
   year={2021},
   publisher={Nature Publishing Group UK London},
   doi={10.1038/s42003-020-01625-6}
-}
-
-@article{andreatta2023tcell,
-  title={T Cell Clonal Analysis Using Single-cell RNA Sequencing and Reference Maps},
-  author={Andreatta, Massimo and Gueguen, Paul and Borcherding, Nicholas and Carmona, Santiago J.},
-  journal={Bio-protocol},
-  year={2023},
-  doi={10.21769/BioProtoc.4735}
 }
 
 @Article{eddelbuettel2011rcpp,
