### 2024-11-06 1726619
fix bib indent and gitignore lintr
