+++ joss/paper.bib
@@ -109,16 +109,16 @@
 }
 
 @article{benzanson2017julia,
-    title={Julia: A fresh approach to numerical computing},
-    author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-    journal={SIAM {R}eview},
-    volume={59},
-    number={1},
-    pages={65--98},
-    year={2017},
-    publisher={SIAM},
-    doi={10.1137/141000671},
-    url={https://epubs.siam.org/doi/10.1137/141000671}
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM {R}eview},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM},
+  doi={10.1137/141000671},
+  url={https://epubs.siam.org/doi/10.1137/141000671}
 }
 
 @article{wang2021single,
