### 2024-11-06 #6868 @Qile0317
@skanwal I removed the duplicate entries just now in the bibliography on my joss branch.
