### 2024-11-06 #6868 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
