+++ joss/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 T and B lymphocytes exhibit diverse surface receptor repertoires that interact with antigens to communicate and protect the body from foreign invaders. [@den2014activation] Modern methods allows the experimental characterization of the collective identity of the immune cells in tissue samples by sequencing RNA transcripts and cell receptors with scRNA-seq and scTCR-seq / scBCR-seq. Such methods generate high-dimensional data which are analyzed with the aid of dimensional reduction techniques to allow annotation of cell subsets. [@huang2022role]
 
-The R [@r2023r] package APackOfTheClones implements and *extends* a novel method to visualize clonal expansion (abbreviated as CE) at a single-cell (abbreviated as SC) resolution using circle packing, along with many clonal analysis utilities. Clonotype frequencies for each cell subset is counted and the values are used as radii and packed into one circular cluster with the largest circles near the center, and overlaid onto a the centroids of each cell subset on the corresponding 2D dimensional reduction plot. The package easily integrates into existing analysis pipelines using the *Seurat* [@hao2023dictionary] and *scRepertoire* [@borcherding2020screpertoire] packages. The original implementation was created in the julia language [@benzanson2017julia] by Christian Murray and Ben Murrell and utilized in two immunology papers. [@ma2021single; @stark2022recombinant]
+The R [@r2023r] package APackOfTheClones implements and *extends* a novel method to visualize clonal expansion (abbreviated as CE) at a single-cell (abbreviated as SC) resolution using circle packing, along with many clonal analysis utilities. Clonotype frequencies for each cell subset is counted and the values are used as radii and packed into one circular cluster with the largest circles near the center, and overlaid onto a the centroids of each cell subset on the corresponding 2D dimensional reduction plot. The package easily integrates into existing analysis pipelines using the *Seurat* [@hao2023dictionary] and *scRepertoire* [@borcherding2020screpertoire] packages. The original implementation was created in the Julia language [@benzanson2017julia] by Christian Murray and Ben Murrell and utilized in two immunology papers. [@ma2021single; @stark2022recombinant]
 
 <!-- should those last two sentences be here or in statement of need? -->
 
@@ -40,7 +40,7 @@ There is no standardized convention to visualize this on the dimensional reducti
 From a visual standpoint, these approaches capture approximate trends that lack precision in depicting the true diversity and abundance of clonal populations for every cell subset. For instance, using a color gradient or contour lines to depict clones fails to directly differentiate between different clonotypes within the same cell subset. This is particularly problematic in scenarios where identifying rare clonotypes is essential, such as detecting the presence of emerging clones in response to a therapy or tracking the evolution of immune responses in different contexts.
 <!-- pretty unsure about this paragraph and whether it captures enough:P -->
 
-APackOfTheClones offers an approach to address these issues by representing exact sizes of each clonotype, in a manner that corresponds exactly to the relevant cell subset. This level of granularity is helpful uncovering hidden patterns, identifying rare clonal populations, and precisely quantifying the impact of therapeutic interventions on immune responses. The original julia implementation's successful utilization in two publication is promising evidence in its potential [@ma2021single; @stark2022recombinant] - however, language choice and the lack of a user-friendly interface limited its accessibility and integration into existing pipelines. The integration of this approach within Seurat and scRepertoire in the popular R ecosystem bridges this gap, increasing accessibility to the wider bioinformatics community through a straightforward, familiar interface.
[…]
