### 2024-11-18 6c59e6b
update citation information for joss
