+++ README.md
@@ -8,7 +8,6 @@
 [![Codecov test coverage](https://codecov.io/gh/Qile0317/APackOfTheClones/branch/main/graph/badge.svg)](https://app.codecov.io/gh/Qile0317/APackOfTheClones?branch=main)
 [![Documentation](https://img.shields.io/badge/docs-stable-blue.svg)](https://qile0317.github.io/APackOfTheClones/)
 [![Developmental Documentation](https://img.shields.io/badge/docs-dev-blue.svg)](https://qile0317.github.io/APackOfTheClones/dev/)
-[![DOI](https://zenodo.org/badge/551699219.svg)](https://doi.org/10.5281/zenodo.13916955)
 [![R Universe APackOfTheClones status badge](https://qile0317.r-universe.dev/badges/APackOfTheClones)](https://qile0317.r-universe.dev/APackOfTheClones)
 [![MIT license](https://img.shields.io/badge/license-MIT-green.svg)](https://github.com/Qile0317/APackOfTheClones/blob/main/LICENSE.md)
 <!-- badges: end -->
@@ -58,39 +57,17 @@ vizAPOTC(pbmc)
 vizAPOTC(pbmc, orig.ident = c("P17B", "P17L"), seurat_clusters = c(1, 3, 4))
 ```
 
-For finer control of the visualization / other parameters, the data for a run can be stored and adjusted with ```RunAPOTC()``` and ```AdjustAPOTC```, then visualized with ```APOTCPlot()```, like so:
-
-```R
-# using the same seurat object as before, run and store associated data,
-# identifying it with a run id. If not, one will be generated.
-pbmc <- RunAPOTC(pbmc, run_id = "default_run_1")
-
-# plot the corresponding data with the run_id. If not provided, will use the latest run.
-APOTCPlot(pbmc)
-
-# adjust the sizes of the clones and move the 2nd cluster to (0, 0). Run id this time was
-# left empty for no particular reason, but it will use "default_run_1" since its the latest
-pbmc <- AdjustAPOTC(
-    pbmc,
-    new_clone_scale_factor = 0.25,
-    relocate_cluster = 2,
-    relocation_coord = c(0, 0)
-)
-
-# one can also do another run with different parameters and a different id.
-pbmc <- RunAPOTC(pbmc, run_id = "run2", try_place = TRUE, repulse = FALSE)
-
-# plot both runs side by side
-library(cowplot)
-cowplot::plot_grid(
-    APOTCPlot(pbmc, run_id = "default_run_1"), APOTCPlot(pbmc, run_id = "run2")
-)
-
-```
+Additional features and use-cases are covered in the vignettes and documentation, including options to store seperate runs with different parameters on the same seurat object, highlighting clonotypes, getting shared clonotypes across clusters, displaying links between them, and more.
 
 ### Package conventions
 
 Most exported functions are named with `camelCase` with the exception of those that modify Seurat objects having `PascalCase` to mimic their conventions. All function arguments follow `snake_case`.
+
+## Please Cite
+
+If you use `APackOfTheClones` in your work, please cite the corresponding [publication](https://joss.theoj.org/papers/eb9d2b2fc3ffb1d440bd7120b66f1f71) with the output of `citation("APackOfTheClones")`:
+
+> Yang, Q., (2024). APackOfTheClones: Visualization of clonal expansion with circle packing. Journal of Open Source Software, 9(103), 6868, https://doi.org/10.21105/joss.06868
 
 ## Documentation
 
@@ -109,11 +86,9 @@ All exported functions have function level documentation.
 
[…]
