+++ cran-comments.md
@@ -3,3 +3,4 @@
 0 errors | 0 warnings | 1 note
 
 * The note generated shouldn't be an issue and has been previously ignored by CRAN maintainers.
+* The previous issue was fixed by buildignoring a file irrelevant to the package itself.
