### 2024-11-18 84f8518
Increment version number to 1.2.3
