+++ cran-comments.md
@@ -2,5 +2,5 @@
 
 0 errors | 0 warnings | 1 note
 
-* The note generated shouldn't be an issue and has been previously ignored by CRAN maintainers.
-* The previous issue was fixed by buildignoring a file irrelevant to the package itself.
+* The note about size generated shouldn't be an issue and has been previously ignored by CRAN maintainers.
+* The previous issues were fixed by removing/buildignoring irrelevant files at build time.
