+++ .github/workflows/R-CMD-check.yaml
@@ -49,4 +49,5 @@ jobs:
 
       - uses: r-lib/actions/check-r-package@v2
         with:
+          args: 'c("--as-cran")'
           upload-snapshots: true
