+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: APackOfTheClones
 Type: Package
 Title: Visualization of Clonal Expansion for Single Cell Immune Profiles
-Version: 1.2.3
+Version: 1.2.4
 Authors@R: c(
     person(
         "Qile",
