+++ cran-comments.md
@@ -1,6 +1,0 @@
-## R CMD check results
-
-0 errors | 0 warnings | 1 note
-
-* The note about size generated shouldn't be an issue and has been previously ignored by CRAN maintainers.
-* The previous issues were fixed by removing/buildignoring irrelevant files at build time.
