+++ cran-comments.md
@@ -1,5 +1,0 @@
-## R CMD check results
-
-0 errors | 0 warnings | 1 note
-
-* The note generated shouldn't be an issue and has been previously ignored by CRAN maintainers.
