+++ NEWS.md
@@ -1,3 +1,5 @@
+# APackOfTheClones (development version)
+
 # APackOfTheClones 1.2.4
 
 ## Additions
